+++ build.gradle
@@ -12,6 +12,7 @@ buildscript {
     ext.gson_version = "2.8.5"
     ext.retrofit_version = "2.5.0"
     ext.moshi_version = "1.8.0"
+    ext.okhttp_version = "3.6.0"
 
     repositories {
         mavenLocal()
@@ -83,6 +84,7 @@ dependencies {
     //Retrofit
     compile "com.squareup.retrofit2:retrofit:$retrofit_version"
     compile "com.squareup.retrofit2:adapter-rxjava2:$retrofit_version"
+    compile "com.squareup.okhttp3:logging-interceptor:$okhttp_version"
 
     //RxKotlin and RxKotlinFX dependencies
     compile "com.github.thomasnield:rxkotlinfx:$rxkotlinfx_version"
