+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/Git4JiraApp.kt
@@ -3,30 +3,30 @@ package com.siliconvalleyoffice.git4jira
 import com.siliconvalleyoffice.git4jira.service.git.GitAuthInterceptor
 import com.siliconvalleyoffice.git4jira.dagger.Injector
 import com.siliconvalleyoffice.git4jira.service.*
+import com.siliconvalleyoffice.git4jira.view.HomeView
 import tornadofx.*
 import view.Git4JiraCredentialsView
 import javax.inject.Inject
 
-class Git4JiraApp : App(Git4JiraCredentialsView::class) {
+class Git4JiraApp : App(HomeView::class) {
 
     @Inject
     lateinit var jsonFileService: Service.JsonFiles
-
-    @Inject
-    lateinit var gitAuthInterceptor: GitAuthInterceptor
 
     init {
         Injector.Instance.initialize()
         Injector.Instance.appComponent.inject(this)
 
-        loadUserConfig()
-    }
-
-    private fun loadUserConfig() {
         jsonFileService.retrieveUserConfig()
 
-        //Define Git Credentials
-        gitAuthInterceptor.setCredentials(jsonFileService.getLastSelectedProject()?.gitService?.credentials)
+        verifyGitHubCrendentials()
     }
 
+    private fun verifyGitHubCrendentials() {
+        jsonFileService.getLastSelectedProject()?.gitService?.gitServiceEnum?.service?.authenticate()?.subscribe({
+            println("GitHub Authentication Successful!")
+        }, {
+            println("GitHub Authentication Failed: $it")
+        })
+    }
 }
