+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/dagger/HomeInjection.kt
@@ -3,6 +3,7 @@ package com.siliconvalleyoffice.git4jira.dagger
 import com.siliconvalleyoffice.git4jira.contract.Home
 import com.siliconvalleyoffice.git4jira.controller.HomeController
 import com.siliconvalleyoffice.git4jira.service.Service
+import com.siliconvalleyoffice.git4jira.service.git.GitAuthInterceptor
 import com.siliconvalleyoffice.git4jira.view.HomeView
 import dagger.Module
 import dagger.Provides
@@ -23,5 +24,6 @@ interface HomeSubComponent {
 class HomeModule(private val homeView: HomeView) {
 
     @Provides
-    fun provideHomeController(loginService: Service.Login, jsonFileService: Service.JsonFiles): Home.Controller = HomeController(homeView, loginService, jsonFileService)
+    fun provideHomeController(loginService: Service.Login, jsonFileService: Service.JsonFiles): Home.Controller
+            = HomeController(homeView, loginService, jsonFileService)
 }
