+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/service/Services.kt
@@ -1,6 +1,8 @@
 package com.siliconvalleyoffice.git4jira.service
 
+import com.siliconvalleyoffice.git4jira.model.GitAuthorizations
 import com.siliconvalleyoffice.git4jira.model.User
+import com.siliconvalleyoffice.git4jira.service.git.GitAuthInterceptor
 import io.reactivex.Single
 import retrofit2.http.GET
 
@@ -10,7 +12,7 @@ import retrofit2.http.GET
  */
 interface GitService : Service.BaseService {
 
-    fun authenticate(): Single<String>
+    fun authenticate(): Single<GitAuthorizations>
 
     fun getUser(username: String): Single<User>
 }
