+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/service/git/GitAuthInterceptor.kt
@@ -4,9 +4,10 @@ import com.siliconvalleyoffice.git4jira.constant.AUTHORIZATION_HEADER
 import com.siliconvalleyoffice.git4jira.constant.EMPTY
 import com.siliconvalleyoffice.git4jira.constant.NO_AUTHENTICATION_HEADER
 import com.siliconvalleyoffice.git4jira.model.Credentials
-import okhttp3.Credentials as OkHttpCredentials
 import okhttp3.Interceptor
 import okhttp3.Response
+import okhttp3.ResponseBody
+import okhttp3.Credentials as OkHttpCredentials
 
 class GitAuthInterceptor : Interceptor {
 
@@ -20,15 +21,7 @@ class GitAuthInterceptor : Interceptor {
             authToken?.let { requestBuilder.addHeader(AUTHORIZATION_HEADER, it) }
         }
 
-        val response = chain.proceed(requestBuilder.build())
-        val responseBuilder = response.newBuilder()
-
-        println(response.code())
-        when(response.code()) {
-            401 -> responseBuilder.message("Authentication Error")
-            200 -> responseBuilder.message("Authenticated")
-        }
-        return responseBuilder.build()
+        return chain.proceed(requestBuilder.build())
     }
 
     fun setCredentials(credentials: Credentials?) {
