+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/service/git/GitRepository.kt
@@ -1,6 +1,8 @@
 package com.siliconvalleyoffice.git4jira.service.git
 
+import com.siliconvalleyoffice.git4jira.model.GitAuthorizations
 import com.siliconvalleyoffice.git4jira.model.User
+import com.siliconvalleyoffice.git4jira.util.GITHUB_API_BASE_URL
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Path
@@ -8,8 +10,8 @@ import retrofit2.http.Url
 
 interface GitRepository {
 
-    @GET("/")
-    fun authenticate(): Single<String>
+    @GET()
+    fun authenticate(@Url baseUrl: String = GITHUB_API_BASE_URL): Single<GitAuthorizations>
 
     @GET("users/{user}")
     fun getUser(@Path("user") user: String): Single<User>
